Move the compute-intensive partition migration test to a separate run [HZ-1051]
Based on thread dump analysis and slow operation detector, we can see that `com.hazelcast.internal.util.HashUtil.MurmurHash3_x64_64(HashUtil.java:177)` is a compute-intensive function that is frequently called during the initial partitions' owners' assignments.
To prevent the `testMigrationStats_afterPartitionsLost_when_NO_MIGRATION` timeout failure, we can try to move it to a separate test run that runs independently of the other tests.

Fix https://github.com/hazelcast/hazelcast/issues/19665
